+++ include/dtorr/fs.h
@@ -5,6 +5,6 @@
 
 int init_torrent_files(dtorr_config* config, dtorr_torrent* torrent);
 
-int rw_piece(dtorr_config* config, dtorr_torrent* torrent, unsigned long index, char* buf, unsigned long buf_size, char is_write);
+int rw_piece(dtorr_config* config, dtorr_torrent* torrent, unsigned long long index, char* buf, unsigned long long buf_size, char is_write);
 
 #endif
