+++ src/msg_out.h
@@ -9,7 +9,7 @@ int send_interested_status(dtorr_config* config, dtorr_peer* peer, char interest
 
 int send_choked_status(dtorr_config* config, dtorr_peer* peer, char choked);
 
-int send_have(dtorr_config* config, dtorr_torrent* torrent, unsigned long index);
+int send_have(dtorr_config* config, dtorr_torrent* torrent, unsigned long long index);
 
 int send_request(dtorr_config* config, dtorr_torrent* torrent, dtorr_peer* peer, dtorr_piece_request* req);
