+++ src/piece_ex.h
@@ -4,10 +4,10 @@
 #include "dtorr/structs.h"
 
 int piece_recv(dtorr_config* config, dtorr_torrent* torrent, dtorr_peer* peer,
-  unsigned long index, unsigned long begin, char* piece, unsigned long piece_len);
+  unsigned long long index, unsigned long long begin, char* piece, unsigned long long piece_len);
 
-int queue_request(dtorr_config* config, dtorr_peer* peer, unsigned long index,
-                  unsigned long begin, unsigned long length);
+int queue_request(dtorr_config* config, dtorr_peer* peer, unsigned long long index,
+                  unsigned long long begin, unsigned long long length);
 
 int pieces_send(dtorr_config* config, dtorr_torrent* torrent);
