+++ src/stream.c
@@ -3,8 +3,8 @@
 
 #define TEMP_BUF_SIZE 24 * 1024
 
-int extract_sock_msg(SOCKET s, char* buf, unsigned long buf_size, unsigned long* msg_len) {
-  unsigned long bytes_avail;
+int extract_sock_msg(SOCKET s, char* buf, unsigned long long buf_size, unsigned long long* msg_len) {
+  unsigned long long bytes_avail;
   unsigned int len_val;
 
   char temp_buf[TEMP_BUF_SIZE];
@@ -32,7 +32,7 @@ int extract_sock_msg(SOCKET s, char* buf, unsigned long buf_size, unsigned long*
   return 0;
 }
 
-int send_sock_msg(SOCKET s, char* buf, unsigned long msg_size) {
+int send_sock_msg(SOCKET s, char* buf, unsigned long long msg_size) {
   char length_buf[4];
 
   uint_to_bigend(length_buf, (unsigned int)msg_size);
